aula 03/10/23

+++ backend/src/controllers/TopicController.ts
@@ -1,8 +1,10 @@
-import { Topic } from './../entities/topic.entity';
 import { ParseIntPipe } from "@nestjs/common";
 import { Body, Controller, Get, Param, Post, Delete, HttpCode } from "@nestjs/common/decorators";
-import { TopicService } from 'src/services/topic.service';
-
+import { Topic } from "src/entities/topic.entity";
+import { TopicService } from "src/services/topic.service";
+import { HttpStatus } from "@nestjs/common";
+import { HttpException } from "@nestjs/common";
+import { Put } from "@nestjs/common/decorators";
 
 
 @Controller('topics')
@@ -15,9 +17,14 @@ export class TopicController {
         return this.service.findAll();
     }
 
-    @Get(':id')
-    findById(@Param('id', ParseIntPipe) id: number): Promise<Topic> {
-        return this.service.findByid(id);
+     @Get(':id')
+    async findById(@Param('id', ParseIntPipe) id: number): Promise<Topic> {
+        const found = this.service.findByid(id);
+
+    if(!found){
+            throw new HttpException('Topic not found', HttpStatus.NOT_FOUND)
+        }
+        return found;
     }
 
     @Post()
@@ -27,7 +34,24 @@ export class TopicController {
 
     @Delete(':id')
     @HttpCode(204)
-    delete(@Param('id', ParseIntPipe) id: number): Promise<void> {
+    async delete(@Param('id', ParseIntPipe) id: number): Promise<void> {
+        const found = this.service.findByid(id);
+
+        if(!found){
+            throw new HttpException('Topic not found', HttpStatus.NOT_FOUND)
+        }
+
         return this.service.delete(id);
     }
+
+    @Put(':id')
+    async update(@Param('id', ParseIntPipe) id: number, @Body() topic: Topic): Promise<Topic> {
+        const found = this.service.findByid(id);
+
+    if(!found){
+            throw new HttpException('Topic not found', HttpStatus.NOT_FOUND)
+        }
+        return this.service.update(id, topic);
+    }
+
 }
